PoC finished: Items system essentially complete fundamentally. Polish needed perhaps. some small QoL changes i imagine. I can only think of potentially incerasing the damage alpha0 duration, and showing what effects an enemy has visually.

+++ Unity Work/Proof of Concepts/Items/Assets/Scripts/Items Scripts/InventoryHandler.cs
@@ -75,7 +75,7 @@ public class InventoryHandler : MonoBehaviour
          GameObject statView = Instantiate(UiStatPrefab, statContainer.position, statContainer.rotation, statContainer);
          statView.transform.Find("Image").GetComponent<Image>().sprite = Stats.GetImage(s);
          statView.transform.Find("StatFlat").GetComponent<Text>().text = s.flatStat.ToString();
-         statView.transform.Find("StatPerc").GetComponent<Text>().text = (s.percentageStat *100).ToString() + "%";
+         statView.transform.Find("StatPerc").GetComponent<Text>().text = ((int)(s.percentageStat*100)).ToString() + "%";
       }
    }
